Shame on you, MONO!
"We can't really enable 16 bytes alignment until the GC supports it. The whole layout/instance size code must be reviewed because we do alignment calculation in terms of the boxed instance, which leads to unexplainable holes at the beginning of an object embedding a simd type. Bug #506144 is an example of this issue."


git-svn-id: https://subversion.assembla.com/svn/sdrsharp/trunk@806 d3cde5d6-82b8-4669-8b62-43054d6781d8